+++ src/components/typeahead-suggestion-title/TypeaheadSuggestionTitle.vue
@@ -1,0 +1,54 @@
+<template>
+	<span class="wvui-typeahead-suggestion__title">
+		<!--eslint-disable-next-line max-len-->
+		{{ titleChunks[ 0 ] }}<em v-if="titleChunks[ 1 ]" class="wvui-typeahead-suggestion__match">{{ titleChunks[ 1 ] }}</em>{{ titleChunks[ 2 ] }}
+	</span>
+</template>
+
+<script lang="ts">
+import Vue, { PropType } from 'vue';
+import { splitStringAtMatch } from './TypeaheadSuggestionTitleUtils';
+
+export default Vue.extend( {
+	name: 'WvuiTypeaheadSuggestionTitle',
+	props: {
+		query: {
+			type: String as PropType<string>,
+			default: ''
+		},
+		title: {
+			type: String as PropType<string>,
+			required: true
+		}
+	},
+	computed: {
+		/**
+		 * Formats title adding highlighted query if it matches.
+		 *
+		 * @return [ string, string, string ]
+		 */
+		titleChunks(): [ string, string, string ] {
+			return splitStringAtMatch( this.query, this.title );
+		}
+	}
+} );
+</script>
+
+<style lang="less">
+@import ( reference ) '@/themes/wikimedia-ui.less';
+
+.wvui-typeahead-suggestion__title {
+	display: block;
+	margin: 0 0 2px 0;
+	color: @color-base;
+	font-size: @font-size-search-suggestion-title;
+	font-weight: @font-weight-bold;
+}
+
+.wvui-typeahead-suggestion__match {
+	// Remove italic
+	font-style: inherit;
+	text-decoration: underline;
+}
+
+</style>